feat: add equipment incident history and add insert to sql file.

+++ src/main/java/fr/efrei/stif/monitor/model/EquipmentRepository.java
@@ -2,5 +2,5 @@ package fr.efrei.stif.monitor.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipmentRepository extends JpaRepository<Equipment, Long> {
+public interface EquipmentRepository extends JpaRepository<Equipment, Integer> {
 }
